Change chart data flow value to be BigDecimal rounded to 2dp

+++ app/data/ChartData.java
@@ -1,11 +1,12 @@
 package data;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 /**
  * Created by Joe on 05/10/2016.
  */
 public class ChartData {
-    public HashMap<Float, Float> dataList = new HashMap<>();
+    public HashMap<Float, BigDecimal> dataList = new HashMap<>();
 
 }
